[path] Add Windows support
