+++ path/example_test.go
@@ -1,3 +1,6 @@
+//go:build !windows
+// +build !windows
+
 package path
 
 // ////////////////////////////////////////////////////////////////////////////////// //
